printful-client/pyproject.toml
[tool.poetry]
name = "Printful-v2"
version = "2.0.0-beta"
description = "A client library for accessing API Documentation v2 | Printful"
authors = []
readme = "README.md"
packages = [
    {include = "printfulv2"},
]
include = ["CHANGELOG.md", "printfulv2/py.typed"]


[tool.poetry.dependencies]
python = "^3.9"
httpx = ">=0.20.0,<0.29.0"
attrs = ">=22.2.0"
python-dateutil = "^2.8.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["F", "I", "UP"]
